+++ lib/html_types.mli
@@ -1549,7 +1549,7 @@ type ol = [ | `Ol ]
 
 type ol_content = [ | `Li of [ | common | `Int_Value ] ]
 
-type ol_content_fun = [ | `Li of [ | common | `Int_Value ] ]
+type ol_content_fun = [ | script_supporting | `Li of [ | common | `Int_Value ] ]
 
 type ol_attrib = [ | common | `Reversed | `Start | `Ol_Type ]
 
@@ -1566,7 +1566,7 @@ type ul = [ | `Ul ]
 
 type ul_content = [ | `Li of [ | li_attrib ] ]
 
-type ul_content_fun = [ | `Li of [ | li_attrib ] ]
+type ul_content_fun = [ | script_supporting | `Li of [ | li_attrib ] ]
 
 type ul_attrib = [ | common ]
 
@@ -2084,7 +2084,7 @@ type table = [ | `Table ]
 
 type table_content = [ | `Tr ]
 
-type table_content_fun = [ | `Tr ]
+type table_content_fun = [ | script_supporting | `Tr ]
 
 type table_attrib = [ | common | `Summary ]
 
@@ -2120,7 +2120,7 @@ type thead = [ | `Thead ]
 
 type thead_content = [ | `Tr ]
 
-type thead_content_fun = [ | `Tr ]
+type thead_content_fun = [ | script_supporting | `Tr ]
 
 type thead_attrib = [ | common ]
 
@@ -2129,7 +2129,7 @@ type tbody = [ | `Tbody ]
 
 type tbody_content = [ | `Tr ]
 
-type tbody_content_fun = [ | `Tr ]
+type tbody_content_fun = [ | script_supporting | `Tr ]
 
 type tbody_attrib = [ | common ]
 
@@ -2138,7 +2138,7 @@ type tfoot = [ | `Tfoot ]
 
 type tfoot_content = [ | `Tr ]
 
-type tfoot_content_fun = [ | `Tr ]
+type tfoot_content_fun = [ | script_supporting | `Tr ]
 
 type tfoot_attrib = [ | common ]
 
@@ -2165,7 +2165,7 @@ type tr = [ | `Tr ]
 
 type tr_content = [ | `Td | `Th ]
 
-type tr_content_fun = [ | `Td | `Th ]
+type tr_content_fun = [ | script_supporting | `Td | `Th ]
 
 type tr_attrib = [ | common ]
 
@@ -2322,7 +2322,7 @@ type select = [ | `Select ]
 
 type select_content = [ | `Optgroup | `Option ]
 
-type select_content_fun = [ | `Optgroup | `Option ]
+type select_content_fun = [ | script_supporting | `Optgroup | `Option ]
 
 type select_attrib =
   [ | common | `Autofocus | `Multiple | `Name | `Size | `Form | `Disabled | `Required
@@ -2343,7 +2343,7 @@ type optgroup = [ | `Optgroup ]
 
 type optgroup_content = [ | `Option ]
 
-type optgroup_content_fun = [ | `Option ]
+type optgroup_content_fun = [ | script_supporting | `Option ]
 
 type optgroup_attrib = [ | common | `Disabled | `Label ]